src/requirements.txt
beautifulsoup4==4.12.3
boto3==1.34.117
botocore==1.34.117
bs4==0.0.2
httpx==0.27.0
huggingface-hub==0.24.5
Jinja2==3.1.4
langchain==0.2.16
langchain-anthropic==0.1.22
langchain-community==0.2.16
langchain-core==0.2.39
langchain-openai==0.1.23
langchain-text-splitters==0.2.2
langfuse==2.33.0
langsmith==0.1.117

openai==1.44.1
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-xlib==0.33
python3-xlib==0.15
qdrant-client==1.11.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
ruff==0.4.6
s3transfer==0.10.1
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.37.2
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
yarl==1.9.4
aioredis
pillow==10.1.0
pandas
colorlog
google-genai